python-rasa-backend/requirements.txt
# RASA
rasa==3.6.13
rasa-sdk==3.6.2

# ML/AI
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
clip-by-openai==1.0

# NLP & Sentiment
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
requests==2.31.0
python-multipart==0.0.6

# Vector DB & Embeddings
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Data Processing
pandas==2.1.3
numpy==1.26.2
pillow==10.1.0

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Security
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
bcrypt==4.1.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
pydantic==2.5.2
python-dateutil==2.8.2
